Current master page: master-1.2.5.php
Never mind what I wrote in the latest email. I thought this newest
version wasn't working properly, but it was just a problem with
responsiveness (I thought the main content wasn't floating properly -
process slider, and recipe details), but the browser window was not the
full size, so it was stacking).